Install react-icons package
Added the react-icons package as a dependency using npm i command. This package provides a collection of popular icons that can be easily used in a React application. The installed package will be used to add icons to various components in the application for enhanced user experience and visual appeal.